Add initial Lambda to replicate RDS Snapshots
To import RDS snapshots to the landing-zone, it should be possible to
export existing snapshots as they are already in Apache Parquet format.

To do this we need to subscribe to an AWS Sns event to run a Lambda
function. This commit gets us to the point of listing the snapshots so
that we can (in subsequent commits) export the latest snapshot to our S3
landing zone.

Co-authored-by: joates-madetech <james.oates@madetech.com>
Co-authored-by: maysakanoni <maysa@madetech.com>
Co-authored-by: b-dalton <ben.dalton@madetech.com>