Add IAM role for rds export to s3
Added a aws provider with new account to deploy the iam role to as this
is the account which contains the snapshot we would like to export. An
IAM tole is created to allow us to use the aws cli to check the export
file is in a parquet format.

 Co-authored-by: joates-madetech <james.oates@madetech.com>
 Co-authored-by: b-dalton <ben.dalton@madetech.com>
 Co-authored-by: mattbee <matt.bee@madetech.com>