Reintroduce google cloud project set up terraform
We have credentials in a Github secret which we will use to auth to create the
required Google Cloud set up to access and import data from Google sheets.

We then write this secret to a file to simplify the auth in terraform.

We can't create projects via terraform, so we now only interact with manually
created Google projects, starting with `dataplatform-stg`

Co-authored-by: joates-madetech <james.oates@madetech.com>
Co-authored-by: maysakanoni <maysa@madetech.com>
Co-authored-by: elena-vi <elena@madetech.com>