Allow AWS assume role block to be dynamic
We want to be able to deploy Terraform locally to test things out on the
DevScratch account.

With `assume_role` included in the provider, we couldn't do this as we
couldn't assume roles using our SSO credentials. If we removed the block
to do local deploys, it would risk breaking staging and production.

This 'work around' (some might say hacky fix) allows the `assume_role`
block to be included dynamically by including a variable of list type
with just one value (arbitrarily set to `true`) to determine the
`assume_role` block inclusion.

By leaving the `assume_roles` variable with no list items, it skips the
`assume_role` block altogether.

Co-authored-by: joates-madetech <james.oates@madetech.com>
Co-authored-by: maysakanoni <maysa@madetech.com>