Increase the diskspace of the RDS instance (#89)
The Liberator dataset requires 5GB, and so 15GB felt like a fair size to accommodate growth.

Also, mark the rds instance as not requiring a final snapshot.  This makes destroying the RDS instance a lot easier with terraform.  If this instance is destroyed and recreated this wouldn't be an issue for the data platform, as this RDS instance is only used to temporarily store data.