Prevent multiple GitHub deploy workflows being executed at once
We've seen workflows fail when main is pushed to in quick succession.

This is because the second terraform process is unable to acquire the
lock.

This is annoying as the second commit won't be deployed, meaning
your fantastic change won't be deployed to the platform.

GitHub Actions have a beta feature called "concurrency" which
aims to fix this deficiency in the workflow executor by queuing
up that second build until the first finishes.

See: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#concurrency