Skip empty tables
If a table in Liberator is missing, or hasn't been updated for a
particular day this shouldn't prevent other tables from being
copied over to the raw zone.

Currently in production the `liberator_hht_singleview_lookups`
table isn't being updated because of a different defect, which
is causing all downstream liberator jobs to fail.

By skiping over the empty dataframe we will prevent the error
"IllegalArgumentException: 'Partition column import_year not found in schema'"

This error happens because an empty dataframe doesn't have
any columns, let alone an import_year.

After this is merged, only jobs which depend on up to date
"liberator_hht_singleview_lookups" data to exist will fail
(which currently is none).