Run the housing repairs crawler on a schedule
Previously this was configured to trigger after all
repairs "Address Matching" jobs had complete.

This wasn't working, and we believe this is due to this
documentation:

>  All jobs or crawlers in a dependency chain must be
>  descendants of a single scheduled or on-demand trigger.

In our case, we have mutliple schedules for each of the different
repair sheets.

We have decided to move to a scheduled based approach as it
will avoid this problem.  Another approach we considered
was to run the crawler at the end of each workflow.  We
were concerned about that erroring if multiple workflows
wanted to run the crawler at the same time.

See also:
https://docs.aws.amazon.com/glue/latest/dg/about-triggers.html

Co-authored-by: b-dalton <ben.dalton@madetech.com>
Co-authored-by: Emma Corbett <emma@madetech.com>