Add push_down_predicate to further filter the partitions to consider
This job was getting very slow > 3 hours to complete, even without
any data to copy.

By adding a push down predicate Spark can reduce the number of partitions
to consider, reducing the duration of the job back down to 15 minutes.

5 days was chosen as a good compromise between speed and catching any
previously missed data in the event of job failure.

Note that this job uses bookmarks to prevent data from being reprocessed.
It would appear that this process of bookmark checking is slow.