A glue job that will ingest all data from one of the tascomi endpoints
The endpoint required can be specified using the resource job parameter.
This commit adds an on demand trigger for the applications endpoint to demonstrate this.

It also adds a workflow for the tascomi data and a crawler for all datasets in the raw zone.

Currently this job will import all data from the endpoint as opposed to just the data we don't currently have.
Hence why the trigger has not been scheduled.