Use docker compose file to run unit tests
In docker compose files you can use relative paths when specifying volumes to mount
The docker cli, however, does not which was causing descrepencies in documentation in how to get the
current working directory in order to mount the scripts and tests.
To help keep it consistent, this commit switches to use docker compose.